min/max are built-ins
